Make the code more palatable for win32.
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/sidplay-libs-2.1.1@23 75748be7-0387-4326-94c7-f41ce0883d2c

+++ builders/resid-builder/src/resid-emu.h
@@ -64,7 +64,7 @@ public:
     static   int  devices (char *error);
 
     // Specific to resid
-    void sampling (uint_least32_t systemclock, uint_least32_t freq);
+    void sampling (float systemclock, float freq);
     bool filter   (const sid_filter_t *filter);
     void model    (sid2_model_t model);
     // Must lock the SID before using the standard functions.
